Add files via upload
Changes made:
*=trim function
*=proper function
*Ensured date and currency format is cohesive across all data and ensured no special characters were listed in our data to eliminate any errors when importing into SQL
*Deleted "prior" column as the reporting needed did not include any data from the "prior" column.
*Deleted column "A" which displayed a list of numbers-almost the same as column "B" (S.No). I decided to delete A instead of B because A began with zero instead of one.

These small and subtle changes made to this dataset has enhanced the quality of data. The data is cohesive throughout the dataset which makes it easier to interpret. The changes I have made lessens the possibility of errors and/or misreading's when importing data into a relational database. We have deleted unnecessary and repetitive columns/data. We can now make reports that will accurately reflect the data.